requirements.txt
timm==0.4.9
transformers==4.12.5
ruamel_yaml
opencv-python
scikit-image
matplotlib
torch==1.7.1
torchvision==0.8.2
pycocotools
pycocoevalcap
streamlit==1.10.0
streamlit_chat==0.0.2
seaborn==0.11.2
mosestokenizer==1.2.1
sentencepiece==0.1.96